Renaming class generator
