+++ generator/dummyClasses/awsDummyClass.js
