+++ generator/dummyClasses/azureDummyClass.js
