+++ generator/generator/Azure/generator.ts
