+++ generator/node-cloud.yml
