+++ generator/package.json
