+++ generator/parser/AWS/parser.ts
