+++ generator/parser/Azure/parser.ts
