+++ generator/transformer/AWS/transformer.ts
