+++ generator/transformer/Azure/transformer.ts
