+++ package.json
@@ -3,7 +3,7 @@
   "private": "true",
   "workspaces": [
     "packages/*",
-    "classGenerator"
+    "generator"
   ],
   "description": "⚡️ The Node.js API for open cloud",
   "scripts": {
